requirements.txt
authlib ~= 1.0
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
psycopg2-binary
pytest~=8.2.2
pytest-django
python-dotenv ~= 0.19
requests ~= 2.27
django-cors-headers
python-jose[cryptography]